[subscribe_service] Ограничение доступов обычным юзерам

+++ subscribe_service/api/v1/endpoints/payment.py
@@ -9,7 +9,11 @@ from constants import PaymentStatus, PaymentType
 from schemas.payment import PaymentCreateScheme, PaymentDBScheme, PaymentFindScheme, PaymentUpdateScheme
 from schemas.user import User
 from security import security_jwt
+from schemas.user import User
 from services.payment import PaymentService, get_payment_service
+from fastapi.exceptions import HTTPException
+from starlette import status
+from schemas.user import Rules
 
 
 router = APIRouter()
@@ -23,6 +27,9 @@ async def create_payment(
 ) -> PaymentDBScheme | None:
     """Создание оплаты."""
 
+    if Rules.admin_rules not in user.rules:
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN)
+
     return await payment_service.create(
         **payment_create.dict(),
         user_id=user.id,
@@ -33,22 +40,33 @@ async def create_payment(
 async def get_payment_by_id(
     id: UUID,
     payment_service: Annotated[PaymentService, Depends(get_payment_service)],
-    user: Annotated[dict | None, Depends(security_jwt)],
+    user: Annotated[User | None, Depends(security_jwt)],
 ) -> PaymentDBScheme:
-    return await payment_service.get(id)
+    """Получение платежа по его ID."""
+
+    payment = await payment_service.get(id)
+    if payment.user_id == user.id or Rules.admin_rules in user.rules:
+        return payment
+
+    raise HTTPException(status_code=status.HTTP_403_FORBIDDEN)
 
 
 @router.get("/", response_model=list[PaymentDBScheme])
 async def get_payments(
     paginate: Annotated[PaginateQueryParams, Depends(PaginateQueryParams)],
     payment_service: Annotated[PaymentService, Depends(get_payment_service)],
-    user: Annotated[dict | None, Depends(security_jwt)],
+    user: Annotated[User | None, Depends(security_jwt)],
     user_id: Annotated[UUID | None, Query(description="ID пользователя")] = None,
     subscribe_id: Annotated[UUID | None, Query(description="ID подписки")] = None,
     payment_type: Annotated[PaymentType | None, Query(description="Тип подписки")] = None,
     payment_date: Annotated[datetime | None, Query(description="Дата оплаты")] = None,
     status: Annotated[PaymentStatus | None, Query(description="Статус платежа")] = None,
 ) -> list[PaymentDBScheme]:
+    """Получение отфильтрованного списка платежей."""
+
+    if Rules.admin_rules not in user.rules:
+        user_id = user.id
+
     return await payment_service.find(
         PaymentFindScheme(
             user_id=user_id,
@@ -68,11 +86,17 @@ async def get_payments(
 )
 async def update_payment(
     id: UUID,
-    payment: PaymentUpdateScheme,
+    payment_update: PaymentUpdateScheme,
     payment_service: Annotated[PaymentService, Depends(get_payment_service)],
-    user: Annotated[dict | None, Depends(security_jwt)],
+    user: Annotated[User | None, Depends(security_jwt)],
 ) -> PaymentDBScheme | None:
-    return await payment_service.update(id, payment)
+    """Изменение платежа."""
+
+    payment = await payment_service.get(id)
+    if payment.user_id == user.id or Rules.admin_rules in user.rules:
+        return await payment_service.update(id, payment_update)
+
+    raise HTTPException(status_code=status.HTTP_403_FORBIDDEN)
 
 
 @router.delete(
@@ -82,6 +106,12 @@ async def update_payment(
 async def delete_payment(
     id: UUID,
     payment_service: Annotated[PaymentService, Depends(get_payment_service)],
-    user: Annotated[dict | None, Depends(security_jwt)],
+    user: Annotated[User | None, Depends(security_jwt)],
 ) -> PaymentDBScheme | None:
-    return await payment_service.delete(id)
+    """Удаление платежа."""
+
+    payment = await payment_service.get(id)
+    if payment.user_id == user.id or Rules.admin_rules in user.rules:
+        return await payment_service.delete(id)
+
+    raise HTTPException(status_code=status.HTTP_403_FORBIDDEN)
